Move ReportFeedbackButton into own file
